Changed IngredientShelve
Roughly speaking, what was IngredientShelve before is now
IngredientShelve['dict']. This is in preparation for creating
drag-and-drop functionality to the ingredient editor. The plan is to
create IngredientShelve['order'], which will be a list of the ingredient
indices, in the order the ingredients are displayed. This may not be the
best way to go about things, so I might decide to change the way I
implement this.