x/y variables removed when deselecting, changed 2d projection
When deselecting an ingredient or other restriction causes the x or y
variable to be removed from the current set of restrictions, the recipe
is now updated to reflect this. Also, the 2-d projection of variables of
different types is scaled so that width and height of the region are the
same.